% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{canavg}
\alias{canavg}
\title{canavg
This function effectively gets down with its bad self
Perchance.}
\usage{
canavg(
  gla_detrep = NA,
  eyey = -8,
  eyez = 10,
  survmeth = "card",
  weightscalc = TRUE,
  diagnostics = TRUE
)
}
\arguments{
\item{gla_detrep}{raw gla detailed report output}
}
\description{
canavg
This function effectively gets down with its bad self
Perchance.
}
