% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{canavg}
\alias{canavg}
\title{Compute Sky Points and Extract Insolation Values}
\usage{
canavg(
  gla_detrep = NA,
  eyey = -8,
  eyez = 10,
  survmeth = "card",
  weightscalc = TRUE,
  diagnostics = TRUE
)
}
\arguments{
\item{gla_detrep}{A detailed report generated by the Gap Light Analyzer software. Default is NA.}

\item{eyey}{Integer (inches); The y-coordinate of the eye position. Default is -8.}

\item{eyez}{Integer (inches); The z-coordinate of the eye position. Default is 10.}

\item{survmeth}{String; The survey method. Default is "card".}

\item{weightscalc}{Logical; if TRUE, computes weights for averaging. Default is TRUE.}

\item{diagnostics}{Logical; if TRUE, prints three figures: (1) interactive figure of XYZ coordinates of all engraved points on mirror surface; (2) insolation values by sky region as contained in Gap Light Analyzer detailed report; (3) and azimuth/altitude coordinates of all sky points. Default is TRUE.}
}
\value{
A list containing:
\describe{
  \item{direction_weights_vector}{A numeric vector of direction weights. May be NA.}
  \item{full_points_data_tibble}{A tibble containing the full points data.}
  \item{Raw_GLA_report}{The raw report data from the Gap Light Analyzer. May be NA.}
  \item{README}{A string noting whether or not a GLA detailed report was passed to the function and whether or not weighted averaging coefficients were calculated. Useful mainly when running the function multiple times with different argument settings.}
}
}
\description{
This function serves two purposes:
(1) Computes points in the sky associated with all survey points on the surface of a concave hemispherical mirror densiometer.
(2) Extracts insolation values for those points from a 'detailed report' generated by the software Gap Light Analyzer and returns coefficients for weighted averaging of each of four separate measurements.
}
\details{
`canavg` computes points in the sky associated with all survey points on the surface of a concave hemispherical mirror densiometer and optionally extracts insolation values for those points from a 'detailed report' generated by the software Gap Light Analyzer. If a GLA detailed report is passed to this function, it returns coefficients for weighted averaging of each of four separate measurements.
}
\examples{
# Example usage of the canavg function
canavg(gla_detrep = my_report, eyey = -8, eyez = 10, survmeth = "card", weightscalc = TRUE, diagnostics = TRUE)
}
