% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{canavg}
\alias{canavg}
\title{Compute Sky Points and Extract Insolation Values}
\usage{
canavg(
  gla_detrep = NA,
  eyey = -8,
  eyez = 10,
  survmeth = "card",
  weightscalc = TRUE,
  diagnostics = TRUE
)
}
\arguments{
\item{gla_detrep}{A detailed report generated by the Gap Light Analyzer software. Default is NA.}

\item{eyey}{The y-coordinate of the eye position. Default is -8.}

\item{eyez}{The z-coordinate of the eye position. Default is 10.}

\item{survmeth}{The survey method. Default is "card".}

\item{weightscalc}{Logical; if TRUE, computes weights for averaging. Default is TRUE.}

\item{diagnostics}{Logical; if TRUE, prints diagnostic information. Default is TRUE.}
}
\value{
A list of coefficients for weighted averaging of each of four separate measurements.
}
\description{
This function serves two purposes:
1. Computes points in the sky associated with all survey points on the surface of a concave hemispherical mirror densiometer.
2. Extracts insolation values for those points from a 'detailed report' generated by the software Gap Light Analyzer and returns coefficients for weighted averaging of each of four separate measurements.
}
\details{
`canavg` computes points in the sky associated with all survey points on the surface of a concave hemispherical mirror densiometer and extracts insolation values for those points from a 'detailed report' generated by the software Gap Light Analyzer. It returns coefficients for weighted averaging of each of four separate measurements.
}
\examples{
# Example usage of the canavg function
canavg(gla_detrep = my_report, eyey = -8, eyez = 10, survmeth = "card", weightscalc = TRUE, diagnostics = TRUE)
}
