% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{canavg}
\alias{canavg}
\title{Compute coefficients for weighted averaging of multiple canopy densiometer readings taken at a single point on the ground surface.}
\usage{
canavg(
  gla_detrep = NA,
  eyey = -8,
  eyez = 10,
  survmeth = "card",
  weightscalc = TRUE,
  diagnostics = TRUE
)
}
\arguments{
\item{gla_detrep}{Filepath as character string; Raw, unedited detailed report in .txt format generated by GLA. Default is NA.}

\item{eyey}{Integer (inches); Y-coordinate of the observer's eye. Default is -8.}

\item{eyez}{Integer (inches); Z-coordinate of the observer's eye. Default is +10.}

\item{survmeth}{Character string; Survey method. Only "card" is supported at this time. "Card," short for "cardinal directions", indicates that measurements were taken facing N, E, S, and W from any single point on the ground surface. Optionally, can be set to "rose" but associated functionality is still under development.}

\item{weightscalc}{Logical; If TRUE, computes weighting coefficients for averaging. Default is TRUE.}

\item{diagnostics}{Logical; if TRUE, prints three figures: (1) XYZ coordinates of all points in physical engraved grid on mirror surface (interactive figure); (2) insolation values by sky region as contained in Gap Light Analyzer detailed report; (3) and azimuth and altitude coordinates of all sky points. If you are executing function canavg many times automatically, as in a loop, set this to FALSE in order to avoid having many plots appearing on your screen. Default is TRUE.}
}
\value{
A list containing:
\describe{
  \item{direction_weights_vector}{Numeric vector of direction weighting coefficients. Will be NA if arguemnt gla_detrep or argument weightscalc are set to NA.}
  \item{full_points_data_tibble}{Tibble containing the full sky points data. Some columns will remain unpopulated if arguemnt gla_detrep or argument weightscalc are set to NA}
  \item{Raw_GLA_report}{Raw report data from the Gap Light Analyzer. Will be NA if no GLA detailed report is passed to function.}
  \item{README}{Character string noting whether or not a GLA detailed report was passed to the function and whether or not weighted averaging coefficients were calculated. Useful for bookkeeping when running the function multiple times with different argument settings.}
}
}
\description{
Compute coefficients for weighted averaging of multiple canopy densiometer readings taken at a single point on the ground surface.
}
\details{
PACKAGE OVERVIEW --------------------------------------------------------------------------------

`canavg` (contraction of 'canopy average') is principally used to compute coefficients for the weighted averaging of multiple forest canopy cover measurements taken from a single point using a concave spherical mirror canopy densiometer. This is accomplished in four sequential steps.

> Step 1: Calculate the spherical coordinates of points in the sky corresponding to each point on the mirror surface when facing in every direction in which measurements were taken (e.g., there are 96 points on the mirror surface, so there are 96*4 or 384 sky points between N-, E-, S-, and W-facing measurements). These coordinates are expressed in degrees clockwise from due north (azimuth) and degrees above horizontal (altitude). Essentially, these sky points represent points in the sky from which light must originate in order to reflect off the mirror surface at a given point in the correct direction to enter the observer's eye.

> Step 2: Associate each of these sky points with a single insolation value, computed externally using the Gap Light Analyzer (GLA) software. At the conclusion of this step, each sky point has a pair of spherical coordinates and an insolation value. Insolation values may be expressed in any of the units available in GLA (e.g., moles photons/meter^2/day). The choice of insolation units does not matter in this context.

> Step 3: Sum the insolation values of all sky points corresponding to each separate directional densiometer measurement.

> Step 4: Compute the coefficients for weighted averaging. The coefficient for any single direction represents the fraction of total insolation which is contributed by sky points associated only with that single directional canopy densiometer measurement. For example, if a researcher is taking separate densiometer measurements facing N, E, S, and W from a point on the land surface, the coefficient for weighted averaging of the north-facing measurement will be equal to the sum of insolation values of all northern sky points divided by the sum of insolation values of all sky points. Coefficients will sum to 1.

Note that steps 2, 3, and 4 can optionally be suppressed by setting argument weightscalc to FALSE. In this case, the only output of interest is a tibble containing spherical coordinates of sky points. Sky point coordinates can also be computed without passing a GLA detailed report to the function by keeping argument gla_detrep set as NA.

GAP LIGHT ANALYZER (EXTERNAL SOFTWARE INSTRUCTIONS) ---------------------------------------------

GLA is designed to estimate forest understory light environment by using global solar radiation models and hemispherical images of canopy coverage recorded using a DSLR and a fisheye lens. In this context, GLA is being used to generate a detailed report which containst estimated insolation values as would be observed in the absence of forest canopy (i.e. values affected by latitude and climatic parameters like average cloud cover and atmospheric haze).

Use of this function for weighting coefficients calculation is dependent the user having previously used GLA to generate a detailed report containing estimates of insolation values for each of 4050 small sky regions at a user-defined latitude, over either a full year or a user-defined fraction of the year (i.e., the growing season), and with reference to either GLA default or user-defined climatic parameters which affect insolation values (atmospheric haziness, cloudiness, etc.). A detailed report can be generated by using GLA to analyze a blank .bmp file. When passing a blank .bmp file to GLA, no understory insolation values or canopy structure values provided in the detailed report will be meaningful, but above-canopy insolation values will be meaningful. Instructions for how to generate a detailed report in GLA are provided below:

> Step 1: Download Gap Light Analyzer 2.0 (link: https://www.caryinstitute.org/science/our-scientists/dr-charles-d-canham/gap-light-analyzer-gla).

> Step 2: Open MS paint or a similar program and save a blank white image as a bitmap (.bmp) file.

> Step 3: Open GLA.

> Step 4: In GLA, open the empty bitmap that you generated in step 2 by clicking File on the menu bar, then Open Image.

> Step 5: Click Configure on the menu bar, then click Edit Configuration. This opens the Configuration Settings dialog box.

> Step 6: In the Configuration Settings dialog box, under the Site tab, input geographic coordinates and elevation of your study site. Do not modify defulat selections in the Orientation or Topographic Shading sections.

> Step 7: In the Configuration Settings dialog box, under the Resolution tab, set the Number of Azimuth Regions to 90 and the Number of Zenith Regions to 45. Function canavg will not be able to parse the final report if this step is not carried out.

> Step 8: In the Configuration Settings dialog box, under the Resolution tab, set the Growing Season Start Date and the Growing Season end date. These values may be either true growing season beginning and end dates or may be set to January 1st and December 31st, respectively, to estimate insolation over a full year. Leave Solar Time Step as 2 minutes.

> Step 9: In the Configuration Settings dialog box, under the radiation tab, make any adjustments required to Model Parameters and Sky-Region Brightness. Choice of appropriate parameter values (Solar Constant, Cloudiness, Spectral Fraction, etc.) is highly location-dependent. This function documentation does not provide advice or recommendations regarding choice of values during this step. Please refer to GLA documentation at https://rem-main.rem.sfu.ca/downloads/Forestry/GLAV2UsersManual.pdf and local climate data for appropriate Model Parameter selection.

> Step 10: Click Configure on the menu bar, then Register Image. Use the mouse to draw a registration circle over the bitmap. When using GLA as intended, this step is required to both define north within the photo and to remove areas of the image which are outside of the circular photograph area from analysis. Because you are not using a real photograph, how you register the image does not matter. Set the circle, then click Apply and then click Okay on the Image Registration dialog box.

> Step 11: Click Image on the menu bar, then Threshold. Click OK on the Threshold dialog box that appears. When using GLA as intended, this step is required to categorize image pixels as either vegetation or sky. Because you are not using a real photograph, threshold value chosen during step does not matter.

> Step 12: Click Calculate on the menu bar, then Run Calculations. On the Calculations dialog box which appears, ensure that option Log Details to File is checked - this is the option that indicates to GLA that it should generate the detailed report which you will pass to function canavg. When you check this option, a dialog box which allows you to choose a file location and name for the detailed report will appear. Choose the location and file name and press Save. On the Calculations dialog box, ensure that option Canopy Structure and Transmitted Gap Light is checked. Finally, click Calculate.

> Step 13: Close GLA. The detailed report containing insolation values for each of 4050 sky regions will be found at the location and filename which you selected in step 12. This report will be passed to function canavg in R.

CONCAVE SPHERICAL CANOPY DENSIOMETER ------------------------------------------------------------

The canopy densiometer to which this package refers is the 'concave spherical crown densiometer' sold primarily by Forestry Suppliers (Jackson, Mississippi) and widely used by foresters and other field biologists (Manufacturer Parts Number: 43888). This package does not currently support computation of sky point spherical coordinates for the convex mirror densiometer (Manufacturer Parts Number: 43887).

Product purchase link: https://www.forestry-suppliers.com/p/43888/13971/forestry-suppliers-spherical-crown-densiometers

Instructions manual link: https://www.forestry-suppliers.com/Documents/4283_msds.pdf

THREE-DIMENSIONAL SPACE FOR MODELING ------------------------------------------------------------

A three-dimensional model of the spherical densiometer mirror surface is created within function canavg to allow for computation of sky point spherical coordinates. When the densiometer is held with the mirror surface facing directly upwards, the three-dimensional space is arranged as follows:

> The origin (X=0, Y=0, Z=0) is at the center of the imaginary sphere of which the physical mirror surface is a section. This sphere has a radius of 3 inches, so the origin is three inches above the center of the engraved grid on the mirror surface. The mirror surface is the bottom of this otherwise imaginary sphere.

> The X and Y dimensions are parallel to the ground surface:

>> The Y-axis extends to positive infinity in the direction that the observer is facing and to negative infinity behind them.

>> The X-axis is perpendicular to the Y-axis and parallel to the ground surface, extending to positive infinity to the observer's right and to negative infinity to the observer's left.

> The Z-axis extends to positive infinity above the observer and to negative infinity below them

The observer's eye position is mapped in this this three-dimensional space, with a default X coordinate of 0, Y coordinate of -8, and Z coordinate of 10. It is recommended that you measure distances between your eye and the sphere's center to ensure that these defaults are appropriate.
}
\examples{
# Example usage of the canavg function
canavg(gla_detrep = my_report, eyey = -8, eyez = 10, survmeth = "card", weightscalc = TRUE, diagnostics = TRUE)
}
