% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{canavg}
\alias{canavg}
\title{canavg
This function serves two purposes:
1) Computes points in the sky associated with all survey points on the surface of a concave hemispherical mirror densiometer;
2) Extracts insolation values for those points from a 'detailed report' generated by the software Gap Light Analyzer and returns coefficients for weighted averaging of each of four separate measurements.
3) testing testing ...}
\usage{
canavg(
  gla_detrep = NA,
  eyey = -8,
  eyez = 10,
  survmeth = "card",
  weightscalc = TRUE,
  diagnostics = TRUE
)
}
\arguments{
\item{gla_detrep}{Raw gla detailed report output, written as "filepath/filename.txt"}

\item{eyey}{The Y coordinate of the observer's eye when viewing the densiometer as expressed relative to an origin at XYZ coordinates 0, 0, 0. Note - the origin for this XYZ scheme is in the middle of the sphere which is partially described by the physical mirror, exactly 3 inches above (i.e. skyward) of the center of the engraved grid. Defaults to -8.}

\item{eyez}{The Z coordinate of the observer's eye when viewing the densiometer as expressed relative to an origin at XYZ coordinates 0, 0, 0. Note - the origin for this XYZ scheme is in the middle of the sphere which is partially described by the physical mirror, exactly 3 inches above (i.e. skyward) of the center of the engraved grid. Defaults to +10.}

\item{survmeth}{Input either "card" or "rose" (under development). The canopy densiometer may be used in either of two ways - four measurements taken in each of four cardinal directions, or four measurements taken in each of eight "compass rose" directions (i.e. N, NE, E, SE, S, SW, W, NW) using a reduced mirror area.}

\item{weightscalc}{Do you want to calculate coefficients for weighted averaging of four or eight measurements, using insolation values derived from the Gap Light Analyzer detailed report. If you using function canavg to calculate sky points only, then set weightscalc == FALSE.}

\item{diagnostics}{Do you want to generate graphics to assess functionality of the function?}
}
\description{
canavg
This function serves two purposes:
1) Computes points in the sky associated with all survey points on the surface of a concave hemispherical mirror densiometer;
2) Extracts insolation values for those points from a 'detailed report' generated by the software Gap Light Analyzer and returns coefficients for weighted averaging of each of four separate measurements.
3) testing testing ...
}
